renames types folder to @types
